logic for user content protection, other users can delete/edit someones project

+++ controllers/projectController.js
@@ -8,13 +8,13 @@ exports.getAllProjects = async (req, res) => {
 exports.getProjectById = async(req, res) => {
     const { id } = req.params;
     const project = await db.projects.getProjectById(id);
+    console.log(project)
     if (!project) {
         res.status(404).render("404", {
             message: "Project not Found!"
         })
     }
 
-    console.log(project)
     res.render("project_single", { project })
 }
 
@@ -49,26 +49,51 @@ exports.updateProjectForm = async (req,res) => {
         })
     }
     
-    const categories = await db.projects.getAllCategories();
+    const categories = await db.categories.getAllCategories();
     res.render("update_project", { categories, project });
 }
 
 // will handle project update
 exports.updateProject = async (req, res) => {
     const { id } = req.params;
-    console.log(req.body);
+    const project = await db.projects.getProjectById(id);
+
+    let current_user_id = null;
+
+    if (req.session.user) {
+        current_user_id = req.session.user.id;
+    }
+    
+    if (project.created_by != current_user_id) {
+        res.status(405);
+        return res.redirect(`/projects/project/${id}`);
+    }
+
     await db.projects.updateProject(id, req.body.title, req.body.description, req.body.category_id);
     console.log("Project was updated!")
 
-    res.redirect(`/projects/project/${id}`);
+    return res.redirect(`/projects/project/${id}`);
 
 }
 
 //delete project controller
 exports.deleteProject = async (req,res) => {
     const { id } = req.params;
-    await db.projects.deleteProject(id);
-    console.log("Project was deleted!")
+    const project = await db.projects.getProjectById(id);
 
-    res.redirect("/projects");
+    let current_user_id = null;
+
+    if (req.session.user) {
+        current_user_id = req.session.user.id;
+    }
+    
+    if (project.created_by != current_user_id) {
+        res.status(405);
+        return res.redirect(`/projects/project/${id}`);
+    }
+    
+    await db.projects.deleteProject(id);
+    console.log("Project was deleted!") 
+
+    return res.redirect("/projects");
 }
